requirements2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pypi_0
adios2=2.8.3=mpi_openmpi_py37h718e7a0_3
aiofiles=23.2.1=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.12.1=pypi_0
altair=5.0.1=pypi_0
annotated-types=0.5.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
apex=0.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
argparse=1.4.0=pypi_0
asgiref=3.7.2=pypi_0
async-timeout=4.0.3=pypi_0
asynctest=0.13.0=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
bcrypt=3.2.2=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
blosc=1.21.1=h83bc5f7_3
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2023.08.22=h06a4308_0
cached-property=1.5.2=pypi_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.6=pypi_0
cloudpickle=2.2.1=pypi_0
comm=0.1.4=pypi_0
coolname=2.2.0=pypi_0
cryptography=42.0.2=pypi_0
cucim=23.6.0=pypi_0
cupy-cuda12x=11.6.0=pypi_0
cycler=0.11.0=pypi_0
cython=3.0.5=pypi_0
databricks-cli=0.18.0=pypi_0
datasets=2.13.2=pypi_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dicomweb-client=0.56.2=pypi_0
dill=0.3.6=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
docker=6.1.3=pypi_0
ecdsa=0.18.0=pypi_0
einops=0.6.1=pypi_0
entrypoints=0.4=pypi_0
evaluate=0.4.1=pypi_0
exceptiongroup=1.1.3=pypi_0
expiring-dict=1.1.0=pypi_0
expiringdict=1.2.1=pypi_0
fastapi=0.78.0=pypi_0
fastjsonschema=2.18.1=pypi_0
fastrlock=0.8.1=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.7.1=pypi_0
fire=0.5.0=pypi_0
flask=2.2.5=pypi_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
ftfy=6.1.1=pypi_0
gdown=4.7.3=pypi_0
geojson=3.0.1=pypi_0
geos=3.8.0=he6710b0_0
girder-client=3.1.14=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gradio=3.34.0=pypi_0
gradio-client=0.2.6=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.60.1=pypi_0
gunicorn=20.1.0=pypi_0
h11=0.12.0=pypi_0
hdf5=1.12.2=mpi_openmpi_hb3f3608_0
httpcore=0.15.0=pypi_0
httpx=0.23.0=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=pypi_0
imagecodecs=2021.11.20=pypi_0
imageio=2.31.2=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=5.2.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
itk=5.3.0=pypi_0
itk-core=5.3.0=pypi_0
itk-filtering=5.3.0=pypi_0
itk-io=5.3.0=pypi_0
itk-numerics=5.3.0=pypi_0
itk-registration=5.3.0=pypi_0
itk-segmentation=5.3.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keybert=0.8.4=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.6.12=pypi_0
krb5=1.19.2=hac12032_0
ld_impl_linux-64=2.38=h1181459_1
libadios2=2.8.3=mpi_openmpi_h3072b55_3
libcurl=7.86.0=h91b91d3_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libllvm11=11.1.0=h9e868ea_6
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=12.2.0=h46fd767_19
libzlib=1.2.13=h166bdaf_4
lightning-utilities=0.10.1=pypi_0
linkify-it-py=2.0.3=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.38.0=py37h4ff587b_0
lmdb=1.4.1=pypi_0
lz4-c=1.9.3=h295c915_1
mako=1.2.4=pypi_0
markdown=3.4.4=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mlflow=1.30.1=pypi_0
monai=1.1.0=pypi_0
monailabel=0.6.0=pypi_0
mpi=1.0=openmpi
mpi4py=3.1.4=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.6.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.8=pypi_0
networkx=2.6.3=pypi_0
nibabel=4.0.2=pypi_0
ninja=1.10.2.3=pypi_0
nltk=3.8.1=pypi_0
notebook=6.5.6=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.55.1=py37h51133e4_0
numpy=1.19.5=pypi_0
numpymaxflow=0.0.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.24.0=pypi_0
openai=1.18.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openmpi=4.1.4=ha1ae619_101
openslide-python=1.1.2=pypi_0
openssl=1.1.1w=h7f8727e_0
orjson=3.9.7=pypi_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
passlib=1.7.4=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.17.1=pypi_0
prometheus-flask-exporter=0.23.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.14=pypi_0
pydantic-core=2.14.6=pypi_0
pydicom=2.3.0=pypi_0
pydicom-seg=0.4.0=pypi_0
pydub=0.25.1=pypi_0
pygments=2.16.1=pypi_0
pyjwt=2.8.0=pypi_0
pynetdicom=2.0.2=pypi_0
pynrrd=0.4.3=pypi_0
pyparsing=3.1.1=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.7.15=h7a1cb2a_1
python-dateutil=2.8.2=pypi_0
python-dotenv=0.20.0=pypi_0
python-jose=3.3.0=pypi_0
python-multipart=0.0.5=pypi_0
python_abi=3.7=2_cp37m
pytorch-ignite=0.4.10=pypi_0
pytorch-lightning=1.9.5=pypi_0
pytz=2022.7.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.1=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
responses=0.18.0=pypi_0
retrying=1.3.4=pypi_0
rfc3986=1.5.0=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
runx=0.0.11=pypi_0
safetensors=0.4.0=pypi_0
schedule=1.1.0=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.2=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=65.5.0=py37h06a4308_0
shapely=1.8.2=pypi_0
simpleitk=2.2.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
snappy=1.1.9=h295c915_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlalchemy=1.4.51=pypi_0
sqlite=3.40.0=h5082296_0
sqlparse=0.4.4=pypi_0
starlette=0.19.1=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.6.0=hdb19cb5_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timeloop=1.0.2=pypi_0
timm=0.9.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toolz=0.12.1=pypi_0
torch=1.13.1=pypi_0
torchaudio=0.13.1=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=py37h06a4308_0
traitlets=5.9.0=pypi_0
transformers=4.30.2=pypi_0
typing-extensions=4.7.1=pypi_0
uc-micro-py=1.0.3=pypi_0
urllib3=1.26.18=pypi_0
uvicorn=0.17.6=pypi_0
watchdog=2.1.8=pypi_0
wcwidth=0.2.9=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
websockets=11.0.3=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.9=pypi_0
xmltodict=0.12.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.2.8=h5eee18b_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h2531618_0
zfp=0.5.5=h295c915_6
zipp=3.15.0=pypi_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=ha4553b6_0
